hangar is working now, not perfect but good enough for now.
The hangar is working now, was pretty easy to copy it from KK. It's not perfect yet, e.g. it doesn't care about possible craft orientation so if the hangar has dimensions of 10, 5, 5 and a vessel with 4, 8, 4 (x, y, z in meters) wants to get stored it won't work.
Right now you will get thrown back to the ksc after storing a vessel but all crew will get added to the colony (a vessel can only be stored if all kerbals can be stored), will have to take a look at how the "ksc" can be moved. I know that the "ksc" location can be switched to e.g. the kerbinside locations but the ksc isn't moved which would be good for this mod.
Also because there are no launchpads yet the vessel will be spawned at the location where it was stored (I think it can even be stored while flying, I'll do something against that) but no crew will be added.

the assumed vessel volume is 80% of it's cubic size (length * height * width) to have a better approximation of the real volume.

Refactor facilities and update configurations

Updated `ColonyData.cfg` to enhance `KC_CAB_Facility` and `KCStorageFacility` entries:
- Updated `serializedData` and `lastUpdateTime`
- Changed `baseGroupName` to "KC_CAB"
- Reorganized `resources` section and added new resources
- Added new `KCStorageFacility` entries and replaced outdated ones

Updated `KerbalColonies.dll` binary file.

In `KerbalColonies.csproj`:
- Replaced `KCBuildingProductionFacility.cs` with `KCProductionFacility.cs`

In `Main.cs`:
- Changed `KSPAddon` attribute to `FlightAndKSC`
- Replaced references to `KCBuildingProductionFacility` with `KCProductionFacility`
- Registered `KCHangarFacility`

Removed `KCBuildingProductionFacility.cs`.

In `KCCrewQuarters.cs`:
- Updated `CrewQuartersInColony` to count `maxKerbals`
- Added static method `AddKerbalToColony`

In `KCFacilityBase.cs`:
- Set `baseGroupName` to "KC_CAB" in `Initialize` method

In `KCHangarFacility.cs`:
- Added `KCHangarFacility` class with methods for managing hangar operations

In `KCStorageFacility.cs`:
- Added `scrollPos` variable for UI scroll management

In `KC_CAB_Facility.cs`:
- Replaced references to `KCBuildingProductionFacility` with `KCProductionFacility`

Added new file `KCProductionFacility.cs` with methods for production management.